Move esphome coordinator to separate module
