+++ homeassistant/components/esphome/dashboard.py
@@ -1,15 +1,10 @@
-"""Files to interact with a the ESPHome dashboard."""
+"""Files to interact with an ESPHome dashboard."""
 
 from __future__ import annotations
 
 import asyncio
-from datetime import timedelta
 import logging
 from typing import Any
-
-import aiohttp
-from awesomeversion import AwesomeVersion
-from esphome_dashboard_api import ConfiguredDevice, ESPHomeDashboardAPI
 
 from homeassistant.config_entries import SOURCE_REAUTH, ConfigEntryState
 from homeassistant.const import EVENT_HOMEASSISTANT_STOP
@@ -17,9 +12,9 @@ from homeassistant.core import CALLBACK_TYPE, Event, HomeAssistant, callback
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.singleton import singleton
 from homeassistant.helpers.storage import Store
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
 from .const import DOMAIN
+from .coordinator import ESPHomeDashboardCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -28,8 +23,6 @@ KEY_DASHBOARD_MANAGER = "esphome_dashboard_manager"
 
 STORAGE_KEY = "esphome.dashboard"
 STORAGE_VERSION = 1
-
-MIN_VERSION_SUPPORTS_UPDATE = AwesomeVersion("2023.1.0")
 
 
 async def async_setup(hass: HomeAssistant) -> None:
@@ -58,7 +51,7 @@ class ESPHomeDashboardManager:
         self._hass = hass
         self._store: Store[dict[str, Any]] = Store(hass, STORAGE_VERSION, STORAGE_KEY)
         self._data: dict[str, Any] | None = None
-        self._current_dashboard: ESPHomeDashboard | None = None
+        self._current_dashboard: ESPHomeDashboardCoordinator | None = None
         self._cancel_shutdown: CALLBACK_TYPE | None = None
 
     async def async_setup(self) -> None:
@@ -70,7 +63,7 @@ class ESPHomeDashboardManager:
             )
 
     @callback
-    def async_get(self) -> ESPHomeDashboard | None:
+    def async_get(self) -> ESPHomeDashboardCoordinator | None:
         """Get the current dashboard."""
         return self._current_dashboard
 
@@ -92,7 +85,7 @@ class ESPHomeDashboardManager:
                 self._cancel_shutdown = None
             self._current_dashboard = None
 
-        dashboard = ESPHomeDashboard(
+        dashboard = ESPHomeDashboardCoordinator(
             hass, addon_slug, url, async_get_clientsession(hass)
         )
         await dashboard.async_request_refresh()
@@ -138,7 +131,7 @@ class ESPHomeDashboardManager:
 
 
 @callback
-def async_get_dashboard(hass: HomeAssistant) -> ESPHomeDashboard | None:
+def async_get_dashboard(hass: HomeAssistant) -> ESPHomeDashboardCoordinator | None:
     """Get an instance of the dashboard if set.
 
     This is only safe to call after `async_setup` has been completed.
@@ -157,43 +150,3 @@ async def async_set_dashboard_info(
     """Set the dashboard info."""
     manager = await async_get_or_create_dashboard_manager(hass)
     await manager.async_set_dashboard_info(addon_slug, host, port)
-
-
-class ESPHomeDashboard(DataUpdateCoordinator[dict[str, ConfiguredDevice]]):  # pylint: disable=hass-enforce-coordinator-module
-    """Class to interact with the ESPHome dashboard."""
-
-    def __init__(
-        self,
-        hass: HomeAssistant,
-        addon_slug: str,
-        url: str,
-        session: aiohttp.ClientSession,
-    ) -> None:
-        """Initialize."""
-        super().__init__(
-            hass,
-            _LOGGER,
-            name="ESPHome Dashboard",
-            update_interval=timedelta(minutes=5),
-            always_update=False,
-        )
-        self.addon_slug = addon_slug
-        self.url = url
-        self.api = ESPHomeDashboardAPI(url, session)
-        self.supports_update: bool | None = None
-
-    async def _async_update_data(self) -> dict:
-        """Fetch device data."""
-        devices = await self.api.get_devices()
-        configured_devices = devices["configured"]
-
-        if (
-            self.supports_update is None
-            and configured_devices
-            and (current_version := configured_devices[0].get("current_version"))
-        ):
-            self.supports_update = (
-                AwesomeVersion(current_version) > MIN_VERSION_SUPPORTS_UPDATE
-            )
-
-        return {dev["name"]: dev for dev in configured_devices}
